feat: Transactional 추가 및 null return 으로 변경

+++ src/main/java/com/example/aboutme/service/SpaceService/SpaceServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.aboutme.service.SpaceService;
 
+import com.example.aboutme.apiPayload.code.status.ErrorStatus;
+import com.example.aboutme.apiPayload.exception.GeneralException;
 import com.example.aboutme.app.dto.SpaceRequest;
 import com.example.aboutme.converter.SpaceConverter;
 import com.example.aboutme.domain.Member;
@@ -8,32 +10,35 @@ import com.example.aboutme.repository.SpaceRepository;
 import com.example.aboutme.service.MemberService.MemberService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @AllArgsConstructor
 public class SpaceServiceImpl implements SpaceService {
     private final SpaceRepository spaceRepository;
     private final MemberService memberService;
-    @Override
+
+    @Transactional
     public Space JoinSpace(SpaceRequest.JoinDTO request) {
         Space newSpace = SpaceConverter.toSpace(request);
         return spaceRepository.save(newSpace);
     }
 
-    @Override
+
     public Space readSpace(Long memberId) {
         Member member = memberService.findMember(memberId);
         return spaceRepository.findByMember_Id(memberId);
     }
 
-    @Override
-    public Space deleteSpace(Long memberId) {
+
+    @Transactional
+    public void deleteSpace(Long memberId) {
         Member member = memberService.findMember(memberId);
+        if (!spaceRepository.existsByMember_Id(memberId)) {
+            throw new GeneralException(ErrorStatus.SPACE_NOT_FOUND);
+        }
         Space targetSpace = spaceRepository.findByMember_Id(memberId);
         spaceRepository.delete(targetSpace);
-        return targetSpace;
     }
 }
